+++ packages/playground/src/index.js
@@ -1,3 +1,6 @@
+import 'core-js/modules/es6.array.find';
+import 'url-search-params-polyfill';
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';